+++ model/model_aluno.py
@@ -1,4 +1,5 @@
 import datetime
+from app import calcula_idade,  empty, media
 dicionario = {
     "Alunos": [
         {
@@ -99,29 +100,3 @@ def delete_aluno(idAluno):
             return aluno
     else:
         return "erro: Aluno não encontrado",404
-
-    #funções
-def calcula_idade(data):   
-    data = data.split("-")  
-    ano = int(data[0])      
-    mes = int(data[1])      
-    dia = int(data[2])      
-    data_nascimento = datetime.date(ano, mes, dia) 
-    hoje = datetime.date.today()
-    idade = hoje - data_nascimento
-    anos = idade.days // 365 
-    return anos
-
-def media(nota1, nota2):
-    soma = nota1 + nota2
-    return soma / 2
-
-def empty(texto):                       
-    texto = texto.capitalize()          
-    if texto in dicionario:             
-        if len(dicionario[texto]) > 0:  
-            return False                
-        else:                           
-            return True                 
-    else:
-        return False
